Добавлен utils, mocs; premium-advertisement вынесен в отдельный компонент
