+++ src/components/premium-advertisement.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const PremiumAdvertisement = () => (
+  <div className="place-card__mark">
+    <span>Premium</span>
+  </div>
+);
+
+export default PremiumAdvertisement;
