+++ src/mocs.js
@@ -1,0 +1,54 @@
+const HOTELS_INFO = [
+  {
+    id: `1`,
+    isFavourite: true,
+    isPremium: true,
+    previewImage: `img/apartment-01.jpg`,
+    price: 130,
+    rating: 1,
+    title: `Beautiful & luxurious studio at great location`,
+    type: `apartment`
+  },
+  {
+    id: `2`,
+    isFavourite: true,
+    isPremium: false,
+    previewImage: `img/apartment-02.jpg`,
+    price: 230,
+    rating: 2.8,
+    title: `Beautiful & luxurious studio at great location`,
+    type: `apartment`
+  },
+  {
+    id: `3`,
+    isFavourite: false,
+    isPremium: false,
+    previewImage: `img/apartment-03.jpg`,
+    price: 410,
+    rating: 3.5,
+    title: `Beautiful & luxurious studio at great location`,
+    type: `apartment`
+  },
+  {
+    id: `4`,
+    isFavourite: false,
+    isPremium: true,
+    previewImage: `img/apartment-01.jpg`,
+    price: 90,
+    rating: 4.3,
+    title: `Beautiful & luxurious studio at great location`,
+    type: `apartment`
+  },
+  {
+    id: `5`,
+    isFavourite: false,
+    isPremium: false,
+    previewImage: `img/room.jpg`,
+    price: 120,
+    rating: 4.8,
+    title: `Beautiful & luxurious studio at great location`,
+    type: `apartment`
+  },
+];
+
+export default HOTELS_INFO;
