+++ src/utils .js
@@ -1,0 +1,7 @@
+const countRating = (rating) => {
+  return `${rating / 5 * 100}%`;
+};
+
+export {
+  countRating
+};
